+++ source/extensions/isaacsim.asset.importer.urdf/python/impl/converter.py
@@ -20,7 +20,12 @@ import os
 import shutil
 
 import omni
-from isaacsim.asset.importer.utils.impl import importer_utils, merge_mesh_utils, stage_utils
+from isaacsim.asset.importer.utils.impl import (
+    importer_utils,
+    merge_mesh_utils,
+    stage_utils,
+    urdf_to_mjc_physx_conversion_utils,
+)
 from pxr import Sdf
 
 from .config import URDFImporterConfig
@@ -127,7 +132,7 @@ class URDFImporter:
             importer_utils.collision_from_visuals(self.stage, self.config.collision_type)
 
         importer_utils.enable_self_collision(self.stage, self.config.allow_self_collision)
-
+        urdf_to_mjc_physx_conversion_utils.convert_joints_attributes(self.stage)
         stage_utils.save_stage(self.stage, intermediate_path)  # save the stage to the output path
         self.stage = None
         gc.collect()