+++ source/extensions/isaacsim.asset.importer.utils/python/impl/importer_utils.py
@@ -21,8 +21,9 @@ import logging
 import os
 from collections.abc import Sequence
 
+import usd.schema.newton
 from isaacsim.asset.transformer import AssetTransformerManager, RuleProfile
-from pxr import PhysxSchema, Sdf, Usd, UsdGeom, UsdPhysics
+from pxr import Gf, PhysxSchema, Sdf, Usd, UsdGeom, UsdPhysics
 
 _logger = logging.getLogger(__name__)
 
@@ -35,6 +36,12 @@ MESH_APPROXIMATION_MAP = {
     "Convex Decomposition": UsdPhysics.Tokens.convexDecomposition,
     "Bounding Sphere": UsdPhysics.Tokens.boundingSphere,
     "Bounding Cube": UsdPhysics.Tokens.boundingCube,
+}
+
+PHYSICS_AXIS_MAP = {
+    "X": UsdPhysics.Tokens.rotX,
+    "Y": UsdPhysics.Tokens.rotY,
+    "Z": UsdPhysics.Tokens.rotZ,
 }
 
 
@@ -129,7 +136,9 @@ def enable_self_collision(usd_stage: Usd.Stage, enabled: bool = True) -> int:
     articulation_roots = [
         prim
         for prim in usd_stage.Traverse()
-        if prim.HasAPI(UsdPhysics.ArticulationRootAPI) or prim.HasAPI("PhysicsArticulationRootAPI")
+        if prim.HasAPI(UsdPhysics.ArticulationRootAPI)
+        or prim.HasAPI("PhysicsArticulationRootAPI")
+        or prim.HasAPI("NewtonArticulationRootAPI")
     ]
 
     if len(articulation_roots) == 0:
@@ -145,19 +154,35 @@ def enable_self_collision(usd_stage: Usd.Stage, enabled: bool = True) -> int:
             attr = physx_api.CreateEnabledSelfCollisionsAttr()
         attr.Set(enabled)
 
+        default_prim.ApplyAPI("NewtonArticulationRootAPI")
+        attr = default_prim.GetAttribute("newton:selfCollisionEnabled")
+        if not attr:
+            attr = default_prim.CreateAttribute("newton:selfCollisionEnabled", Sdf.ValueTypeNames.Bool)
+        attr.Set(enabled)
         return 1
 
     for articulation_root in articulation_roots:
         if not articulation_root or not articulation_root.IsValid():
             continue
 
+        if not articulation_root.HasAPI("PhysicsArticulationRootAPI"):
+            articulation_root.ApplyAPI("PhysicsArticulationRootAPI")
+
         if not articulation_root.HasAPI(PhysxSchema.PhysxArticulationAPI):
             PhysxSchema.PhysxArticulationAPI.Apply(articulation_root)
+
+        if not articulation_root.HasAPI("NewtonArticulationRootAPI"):
+            articulation_root.ApplyAPI("NewtonArticulationRootAPI")
 
         physx_api = PhysxSchema.PhysxArticulationAPI(articulation_root)
         attr = physx_api.GetEnabledSelfCollisionsAttr()
         if not attr:
             attr = physx_api.CreateEnabledSelfCollisionsAttr()
+        attr.Set(enabled)
+
+        attr = articulation_root.GetAttribute("newton:selfCollisionEnabled")
+        if not attr:
+            attr = articulation_root.CreateAttribute("newton:selfCollisionEnabled", Sdf.ValueTypeNames.Bool)
         attr.Set(enabled)
 
     return len(articulation_roots)
@@ -289,3 +314,63 @@ def remove_custom_scopes(stage: Usd.Stage) -> None:
     if scope and scope.IsA(UsdGeom.Scope):
         stage.RemovePrim(Sdf.Path(scope_paths))
     return
+
+
+def create_physx_mimic_joint(prim: Usd.Prim) -> None:
+    """Create a mimic joint for a joint.
+
+    Args:
+        prim: prim to create the mimic joint for.
+    """
+    if prim.HasAPI("NewtonMimicAPI"):
+        # Get the mimic relation as a Usd.Rel
+        mimic_rel = prim.GetRelationship("newton:mimicJoint")
+        if mimic_rel is None or not mimic_rel.IsValid():
+            _logger.warning(f"newton:mimicJoint not found or invalid for prim {prim.GetPath()}")
+            return
+
+        target = mimic_rel.GetTargets()[0]
+        if not target:
+            _logger.warning(f"newton:mimicJoint relationship has no target for prim {prim.GetPath()}")
+            return
+
+        # Read newton:mimicCoef1 and newton:mimicCoef0
+        mimic_coef1_attr = prim.GetAttribute("newton:mimicCoef1")
+        mimic_coef0_attr = prim.GetAttribute("newton:mimicCoef0")
+        mimic_coef1 = mimic_coef1_attr.Get() if mimic_coef1_attr and mimic_coef1_attr.IsValid() else None
+        mimic_coef0 = mimic_coef0_attr.Get() if mimic_coef0_attr and mimic_coef0_attr.IsValid() else None
+
+        target_prim = prim.GetStage().GetPrimAtPath(target)
+        if not target_prim:
+            _logger.warning(f"target prim not found for prim {prim.GetPath()}")
+            return
+        if not target_prim.HasAttribute("physics:axis"):
+            _logger.warning(f"target prim does not have physics:axis attribute for prim {prim.GetPath()}")
+            return
+        target_axis = target_prim.GetAttribute("physics:axis").Get().upper()
+
+        axis = prim.GetAttribute("physics:axis").Get().upper() if prim.HasAttribute("physics:axis") else None
+        if axis is None:
+            _logger.warning(f"prim does not have physics:axis attribute for prim {prim.GetPath()}")
+            return
+
+        target_joint = UsdPhysics.Joint(target_prim)
+        # Check if the target joint is flipped, and if so, invert the gearing (mimic_coef1)
+        local_rot1 = target_joint.GetLocalRot1Attr().Get()
+        if local_rot1.GetReal() == -1 or any(v == -1 for v in local_rot1.GetImaginary()):
+            mimic_coef1 = -mimic_coef1
+            _logger.info(f"Inverted gearing for prim {prim.GetPath()} because target joint is flipped")
+
+        # Create the PhysX mimic joint attribute on the current prim
+        physx_mimic_api = PhysxSchema.PhysxMimicJointAPI.Apply(prim, PHYSICS_AXIS_MAP[axis])
+        mimic_rel_physx = physx_mimic_api.CreateReferenceJointRel()
+        mimic_rel_physx.SetTargets([target])
+        if mimic_coef1 is not None:
+            physx_mimic_api.CreateGearingAttr().Set(mimic_coef1)
+        else:
+            _logger.warning(f"newton:mimicCoef1 not found or invalid for prim {prim.GetPath()}")
+        if mimic_coef0 is not None:
+            physx_mimic_api.CreateOffsetAttr().Set(mimic_coef0)
+        else:
+            _logger.warning(f"newton:mimicCoef0 not found or invalid for prim {prim.GetPath()}")
+        physx_mimic_api.CreateReferenceJointAxisAttr().Set(PHYSICS_AXIS_MAP[target_axis])